Only run animation loop if focused
This improves performance when multiple tabs are open, as chrome we
provide animationFrames for any window visible, even if out of focus,
LeapJS only needs to update if focus is given.  This improved by a
noticeable 5 FPS (15 to 20) with unfocused visible windows.

Note that when focus is returned, the focus event is fired before
resuming the animation loop.