Only run the pipeline on animation frames in the browser
This is pretty important, as the pipeline can be heavy and even include
calls for rendering to the GPU.  It does mean that if someone is needs
to do data analysis on deviceFrames directly, they won't have the same
processed frames available to them.  This isn't a problem until its a
problem.