[FIX] Fix Dependabot ecstatic Dependency
Dependabot is complaining about the ecstatic file server, which I believe is only used in `http-server`, which is only used in a vestigial preview command in the Makefile.

Since VS Code now has `live-server` previewing extensions, I'm just removing that command.

Hopefully this gets us back to a clean universe.